+++ common/scripts/definitions/items/grenades.ts
@@ -440,7 +440,7 @@ export function Grenades_Default_Init(grenades:Definitions<GrenadeDef,{}>){
             rig_arms:DefaultFistRig,
             rig_image:GrenadeRig,
             rank:ItemRank.A,
-            push_force_resistence:0,
+            push_force_resistence:0.1,
         },
     )
 }